Added the new nether songs and some 1.4 music
Namely, I made Chrysopoeia underground crimson, Rubedo underground desert, and So Below underground corruption. With this I changed creative4 and creative5 to underground jungle and slime rain to make room for underground crimson and corruption. I also changed the end music to the space music, and I changed piano2 to windy day to make room for this. The last things I added were changing menu3 to the journey's end title screen, and I made pigstep the otherworld version of the underworld.